Major refactor to pet/charm handling
