+++ game/world/managers/objects/ObjectManager.py
@@ -401,10 +401,6 @@ class ObjectManager:
         return False
 
     # override
-    def get_pet(self):
-        return None
-
-    # override
     def is_pet(self):
         return False
